+++ pom.xml
@@ -307,29 +307,12 @@
 		</dependency>
 
 		<!-- google login -->
-		 
-    <dependency>
-            <groupId>org.apache.httpcomponents</groupId>
-            <artifactId>httpclient</artifactId>
-            <version>4.5.3</version>
-        </dependency>
-        
-       <dependency>
-		    <groupId>org.springframework.social</groupId>
-		    <artifactId>spring-social-google</artifactId>
-	            <version>1.0.0.RELEASE</version>
-	    </dependency>
+		<dependency>
+			<groupId>org.springframework.social</groupId>
+			<artifactId>spring-social-google</artifactId>
+			<version>1.0.0.RELEASE</version>
+		</dependency>
 
-
-        <!-- 
-        <dependency>
-            <groupId>com.github.spring-social</groupId>
-            <artifactId>spring-social-google</artifactId>
-            <version>1.1.3</version>
-        </dependency>
-  -->
-		
-	
 		<!-- 파일 업로드 라이브러리 -->
 		<dependency>
 			<groupId>commons-fileupload</groupId>
