--- src/main/java/happy/land/people/dto/AuthInfo.java
@@ -1,23 +1,0 @@
-package happy.land.people.dto;
-
-public class AuthInfo  {
-	 
-    private String clientId;
-    private String clientSecret;
-    
-    
-    public AuthInfo (String clientId, String clientSecret) {
-        this.clientId = clientId;
-        this.clientSecret = clientSecret;
-    }
- 
- 
-    public String getClientId() {
-        return clientId;
-    }
- 
- 
-    public String getClientSecret() {
-        return clientSecret;
-    }
-}
