+++ src/main/java/happy/land/people/mapper/ChoMapper.xml
@@ -5,22 +5,29 @@
 
 <!-- 내DTO에 맞쳐서 수정하셈; -->
 
-<!-- 회원가입 일반 -->
-<insert id="signUp" parameterType="ChoDto">
-	INSERT INTO LPUSER (USER_EMAIL, USER_PASSWORD, USER_NICKNAME,USER_AUTH, USER_EMAILCHK, USER_EMAILKEY) 
-	SELECT #{user_email}, #{user_password}, #{user_nickname}, 'U', 'N' , '' FROM DUAL WHERE NOT EXISTS(SELECT * FROM LPUSER WHERE USER_EMAIL=#{user_email})
+
+
+
+
+
+
+<!-- 회원가입 (이메일) OK-->
+<insert id="signUpU" parameterType="ChoDto">
+	INSERT INTO LPUSER (USER_EMAIL, USER_PASSWORD, USER_NICKNAME, USER_AUTH,USER_EMAILKEY) 
+SELECT #{user_email}, #{user_password}, #{user_nickname}, 'U', '' FROM DUAL WHERE NOT EXISTS(SELECT* FROM LPUSER WHERE USER_EMAIL =#{user_email})
+	
 </insert>
 
-<!-- 회원가입 (네이버) -->
+<!-- 회원가입 (네이버)  OK-->
 <insert id="signUpN" parameterType="ChoDto">
 	INSERT INTO LPUSER (USER_EMAIL,  USER_NICKNAME,USER_AUTH, USER_EMAILCHK, USER_EMAILKEY) 
-	SELECT #{user_email},  #{user_nickname}, 'N', 'Y' , '' FROM DUAL WHERE NOT EXISTS(SELECT * FROM LPUSER WHERE USER_EMAIL=#{user_email})
+	SELECT #{user_email}, #{user_nickname}, 'N', 'Y' , '' FROM DUAL WHERE NOT EXISTS(SELECT * FROM LPUSER WHERE USER_EMAIL=#{user_email})
 </insert>
 
 <!-- 회원가입 (구글)  OK-->
 <insert id="signUpG" parameterType="ChoDto">
-	INSERT INTO LPUSER (USER_EMAIL,  USER_NICKNAME,USER_AUTH, USER_EMAILCHK, USER_EMAILKEY) 
-	SELECT #{user_email},  #{user_nickname}, 'G', 'Y' , '' FROM DUAL WHERE NOT EXISTS(SELECT * FROM LPUSER WHERE USER_EMAIL=#{user_email})
+	INSERT INTO LPUSER (USER_EMAIL, USER_PASSWORD, USER_NICKNAME, USER_AUTH, USER_EMAILKEY) 
+	VALUES(#{user_email}, #{user_password}, #{user_nickname}, 'G', '')
 </insert>
 
 <!-- 이메일 중복체크  -->
@@ -33,18 +40,16 @@
 </select>
 
 
-<!-- api사용자 이메일 중복체크 -->
-<select id="apiEmailDupChk" parameterType="java.lang.String" resultType="ChoDto">
-	SELECT USER_AUTH FROM LPUSER WHERE USER_EMIAL = #{user_email}
-</select>
+
+
 
 
 
 <!-- 로그인 -->
 <select id="login" parameterType="ChoDto" resultType="ChoDto">
-	SELECT USER_EMAIL, USER_PASSWORD, USER_NICKNAME, USER_AUTH, USER_DELFLAG ,USER_EMAILCHK
+	SELECT USER_EMAIL, USER_PASSWORD, USER_NICKNAME, USER_AUTH, USER_EMAILCHK
 	FROM LPUSER
-	WHERE LOWER(USER_EMAIL) = LOWER(#{user_email})
+	WHERE LOWER(USER_EMAIL) = LOWER(#{user_email}) AND USER_AUTH = 'U'  AND USER_DELFLAG ='F' 
 </select>
