+++ src/main/java/happy/land/people/model/cho/ChoServiceImpl.java
@@ -1,4 +1,4 @@
-package happy.land.people.model.user;
+package happy.land.people.model.cho;
 
 import java.io.UnsupportedEncodingException;
 import java.util.Map;
@@ -32,14 +32,19 @@ public class ChoServiceImpl implements IChoService {
 		boolean isc = iChoDao.signUp(dto);
 		System.out.println("==============================================================="+isc);
 		
+		
 		// 만약에 user_auth가 n 이거나 g면 바로 가입 u면 이메일인증 하기
+		
 		String user_auth = dto.getUser_auth();
-		System.out.println(user_auth);
-		if(user_auth.equals("N")) {
+		
+		if(user_auth == "N") {
+			
 			return isc;
-		}else if(user_auth.equals("G")) {
+		}else if(user_auth == "G") {
 			return isc;
 		}else {
+		
+		
 		//authkey 임시 생성 후 dto 같이 담아줌
 		String user_emailkey = new TempKey().getKey(50, false);
 		dto.setUser_emailkey(user_emailkey);
@@ -51,7 +56,7 @@ public class ChoServiceImpl implements IChoService {
 		// 메일 내용 담을 변수(email, authkey만 보낼예정)
 		String mailContent = new StringBuffer().append("<h1>[이메일 인증]</h1>")
 				.append("<p>아래 링크를 클릭하시면 이메일 인증이 완료됩니다.</p>")
-				.append("<a href='http://192.168.10.186:8091/LandPeople/mailConform.do?user_email=")
+				.append("<a href='http://192.168.4.31:8091/LandPeople/mailConfirm.do?user_email=")
 				.append(dto.getUser_email())
 				.append("&authkey=").append(dto.getUser_emailkey())
 				.append("' target='_blank' >이메일 인증 확인</a>").toString();
@@ -175,11 +180,6 @@ public class ChoServiceImpl implements IChoService {
 		
 		
 		return iChoDao.emailAuthChk(user_email);
-	}
-
-	@Override
-	public ChoDto apiEmailDupChk(String user_email) {
-		return iChoDao.apiEmailDupChk(user_email);
 	}
 
 }
