+++ src/main/java/happy/land/people/model/cho/IChoDao.java
@@ -1,10 +1,13 @@
-package happy.land.people.model.user;
+package happy.land.people.model.cho;
 
 import java.util.Map;
 
 import happy.land.people.dto.cho.ChoDto;
 
-public interface IChoService {
+
+public interface IChoDao {
+
+	
 	//회원가입
 	public boolean signUp(ChoDto dto);
 	//로그인
@@ -21,12 +24,12 @@ public interface IChoService {
 	public boolean authkeyUpdate(ChoDto dto);
 	//이메일 인증 상태 변경
 	public boolean authStatusUpdate(String user_email);
+	
 	//이메일로 하는 비밀번호 수정 (비밀번호 찾기)
 	public boolean findPW(ChoDto dto);
 	
 	//비밀번호 찾기 할때 이메일 무슨가입자인지 구별하는 
-		public int emailAuthChk(String user_email);
-		
-		//api사용자 이메일 중복체크
-		public ChoDto apiEmailDupChk(String user_email);
+	public int emailAuthChk(String user_email);
+	
+	
 }
