+++ src/main/webapp/WEB-INF/views/user/deleteAccount.jsp
@@ -1,23 +1,15 @@
 <%@ page language="java" contentType="text/html; charset=UTF-8"
     pageEncoding="UTF-8"%>
+<% request.setCharacterEncoding("UTF-8"); %>
+<% response.setContentType("text/html; charset=UTF-8"); %>
 <!DOCTYPE html>
 <html>
 <head>
 <meta charset="UTF-8">
-<title>회원탈퇴</title>
+<title>회원 탈퇴</title>
+<!-- 회원 탈퇴 안내 페이지 -->
 </head>
 <body>
-
-정말로회원탈퇴하시겠습니까?<br>
-탈퇴약관.....<br>
-
-...
-
-
-<form action="./delflag.do" method="get">
-<input type="submit" value="탈퇴">
-</form>
-
 
 </body>
 </html>
