+++ src/main/webapp/WEB-INF/views/user/infoEmail.jsp
@@ -6,16 +6,10 @@
 <html>
 <head>
 <meta charset="UTF-8">
-<title>이메일 인증 완료</title>
+<title>인증</title>
+<!-- 인증이 잘됬다고 알려주는 페이지 -->
 </head>
 <body>
 
-인증이 완료되었습니다
-
-로그인하러 ㄱㄱ
-
-<form action="./loginPage.do" method="get">
-	<input type="submit" value="로그인">
-</form>
 </body>
 </html>
