+++ src/main/webapp/WEB-INF/views/users/delpage.jsp
@@ -1,0 +1,27 @@
+<%@ page language="java" contentType="text/html; charset=UTF-8"
+    pageEncoding="UTF-8"%>
+<!DOCTYPE html>
+<html>
+<head>
+<meta charset="UTF-8">
+<title>회원탈퇴</title>
+</head>
+<body>
+
+정말로회원탈퇴하시겠습니까?<br>
+탈퇴약관.....<br>
+
+<pre>
+이거 왜 안쓸까?
+막쳐도 알아서 해주는데
+장애인법 어쩌고 그것떄문인가???아 근ㄴ데 왜 앋ㄴ되는거야 타입이 안맞냐 왜
+</pre>
+...
+
+<form action="./delflag.do" method="get">
+<input type="submit" value="탈퇴">
+</form>
+
+
+</body>
+</html>
