+++ src/main/webapp/WEB-INF/views/users/login/delpage.jsp
@@ -6,9 +6,9 @@
 <html>
 <head>
 <meta charset="UTF-8">
-<title></title>
+<title>Insert title here</title>
 </head>
 <body>
-
+dd
 </body>
 </html>
