+++ src/main/webapp/WEB-INF/views/users/login/logout.jsp
@@ -6,9 +6,17 @@
 <html>
 <head>
 <meta charset="UTF-8">
-<title></title>
+<title>로그아웃 임시</title>
 </head>
 <body>
+
+로그아웃 확인할라고 임시로 만든 페이지임
+${session}
+${ldto}
+<form action="./logout.do" method="get">
+
+<input type="submit" value="로그아웃">
+</form>
 
 </body>
 </html>
