+++ src/main/webapp/WEB-INF/views/users/pwforget/pwemailSubmit.jsp
@@ -6,9 +6,14 @@
 <html>
 <head>
 <meta charset="UTF-8">
-<title></title>
+<title>비번찾기 이메일 보냄</title>
 </head>
 <body>
+<h1>이메일 보냈음 비번찾는거 수정후 로그인해줘</h1>
+
+<form action="./loginPage.do" method="get">
+	<input type="submit" value="로그인">
+</form>
 
 </body>
 </html>
