+++ src/main/webapp/WEB-INF/views/users/sign/auth.jsp
@@ -6,9 +6,16 @@
 <html>
 <head>
 <meta charset="UTF-8">
-<title></title>
+<title>이메일 인증 완료</title>
 </head>
 <body>
 
+인증이 완료되었습니다
+
+로그인하러 ㄱㄱ
+
+<form action="./loginPage.do" method="get">
+	<input type="submit" value="로그인">
+</form>
 </body>
 </html>
